src/BigQuery/requirements.txt
google-cloud-bigquery==3.11.4
pandas==2.0.3
numpy==1.24.3
flask==2.3.2
flask-cors==4.0.0
websockets==11.0.3
python-dateutil==2.8.2
requests==2.31.0
db-dtypes==1.1.1
pyarrow==12.0.1
